# docs/merchant/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/merchant/api/v1/payments/": {
            "post": {
                "description": "Creates a new payment with associated wallets for the current merchant.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment creation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_common_restdto.PaymentCreateRestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_common_restdto.PaymentCreateResponseRestDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Merchant not found or error during payment creation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/token/": {
            "get": {
                "description": "Retrieves a list of all active tokens.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "List active tokens",
                "responses": {
                    "200": {
                        "description": "List of active tokens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.TokenListResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/user/me/": {
            "get": {
                "description": "Retrieves the profile of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.UserMeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user or invalid user type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_1stpay_1stpay_internal_transport_rest_common_restdto.PaymentCreateResponseRestDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_common_restdto.PaymentCreateRestDTO": {
            "type": "object",
            "required": [
                "requested_amount"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "requested_amount": {
                    "type": "number"
                }
            }
        },
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.TokenListResponseDTO": {
            "type": "object",
            "properties": {
                "blockchain_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_native": {
                    "type": "boolean"
                },
                "logo": {
                    "$ref": "#/definitions/null.String"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.UserMeResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}